refactor: Update import paths and restructure edge and node generators
Refactored the import paths for edge and node generators to align with
the new module structure. This change includes moving functions from
`matgraphdb.core.nodes` and `matgraphdb.core.edges` to
`matgraphdb.generators.nodes` and `matgraphdb.generators.edges`.

Additionally, created a new `material_store.py` file to encapsulate
the `MaterialStore` class, improving organization and maintainability
of the codebase.

No functionality is affected as the core logic remains intact, but
the restructuring enhances clarity and prepares for future development.
