+++ docs/source/03_api/core/edge_generators.rst
@@ -1,38 +1,38 @@
 Edge Generators
 ========================
 
-- :func:`element_element_neighborsByGroupPeriod <matgraphdb.core.edges.element_element_neighborsByGroupPeriod>` - A function that generates the neighbors of an element by group and period.
+- :func:`element_element_neighborsByGroupPeriod <matgraphdb.generators.edges.element_element_neighborsByGroupPeriod>` - A function that generates the neighbors of an element by group and period.
 
-- :func:`element_element_bonds <matgraphdb.core.edges.element_element_bonds>` - A function that generates the bonds of an element.
+- :func:`element_element_bonds <matgraphdb.generators.edges.element_element_bonds>` - A function that generates the bonds of an element.
 
-- :func:`element_oxiState_canOccur <matgraphdb.core.edges.element_oxiState_canOccur>` - A function that generates the possible oxidation states of an element.
+- :func:`element_oxiState_canOccur <matgraphdb.generators.edges.element_oxiState_canOccur>` - A function that generates the possible oxidation states of an element.
 
-- :func:`material_chemenv_containsSite <matgraphdb.core.edges.material_chemenv_containsSite>` - A function that generates the sites of a material.
+- :func:`material_chemenv_containsSite <matgraphdb.generators.edges.material_chemenv_containsSite>` - A function that generates the sites of a material.
 
-- :func:`material_crystalSystem_has <matgraphdb.core.edges.material_crystalSystem_has>` - A function that generates the crystal system of a material.
+- :func:`material_crystalSystem_has <matgraphdb.generators.edges.material_crystalSystem_has>` - A function that generates the crystal system of a material.
 
-- :func:`material_element_has <matgraphdb.core.edges.material_element_has>` - A function that generates the elements of a material.
+- :func:`material_element_has <matgraphdb.generators.edges.material_element_has>` - A function that generates the elements of a material.
 
-- :func:`material_lattice_has <matgraphdb.core.edges.material_lattice_has>` - A function that generates the lattice of a material.
+- :func:`material_lattice_has <matgraphdb.generators.edges.material_lattice_has>` - A function that generates the lattice of a material.
 
-- :func:`material_spg_has <matgraphdb.core.edges.material_spg_has>` - A function that generates the space group of a material.
+- :func:`material_spg_has <matgraphdb.generators.edges.material_spg_has>` - A function that generates the space group of a material.
 
-- :func:`element_chemenv_canOccur <matgraphdb.core.edges.element_chemenv_canOccur>` - A function that generates the possible oxidation states of an element.
+- :func:`element_chemenv_canOccur <matgraphdb.generators.edges.element_chemenv_canOccur>` - A function that generates the possible oxidation states of an element.
 
-- :func:`spg_crystalSystem_isApart <matgraphdb.core.edges.spg_crystalSystem_isApart>` - A function that generates the crystal system of a material.
+- :func:`spg_crystalSystem_isApart <matgraphdb.generators.edges.spg_crystalSystem_isApart>` - A function that generates the crystal system of a material.
 
 
 
 .. autosummary::
    :toctree: _autosummary
 
-   matgraphdb.core.edges.element_element_neighborsByGroupPeriod
-   matgraphdb.core.edges.element_element_bonds
-   matgraphdb.core.edges.element_oxiState_canOccur
-   matgraphdb.core.edges.material_chemenv_containsSite
-   matgraphdb.core.edges.material_crystalSystem_has
-   matgraphdb.core.edges.material_element_has
-   matgraphdb.core.edges.material_lattice_has
-   matgraphdb.core.edges.material_spg_has
-   matgraphdb.core.edges.element_chemenv_canOccur
-   matgraphdb.core.edges.spg_crystalSystem_isApart
+   matgraphdb.generators.edges.element_element_neighborsByGroupPeriod
+   matgraphdb.generators.edges.element_element_bonds
+   matgraphdb.generators.edges.element_oxiState_canOccur
+   matgraphdb.generators.edges.material_chemenv_containsSite
+   matgraphdb.generators.edges.material_crystalSystem_has
+   matgraphdb.generators.edges.material_element_has
+   matgraphdb.generators.edges.material_lattice_has
+   matgraphdb.generators.edges.material_spg_has
+   matgraphdb.generators.edges.element_chemenv_canOccur
+   matgraphdb.generators.edges.spg_crystalSystem_isApart
