+++ docs/source/03_api/core/node_generators.rst
@@ -2,28 +2,39 @@ Node Generators
 ========================
 
 
-- :func:`element <matgraphdb.core.nodes.generators.element>` - A function that generates the elements of a material.
+- :func:`element <matgraphdb.generators.nodes.element>` - A function that generates the elements of a material.
 
-- :func:`chemenv <matgraphdb.core.nodes.generators.chemenv>` - A function that generates the chemical environments of a material.
+- :func:`chemenv <matgraphdb.generators.nodes.chemenv>` - A function that generates the chemical environments of a material.
 
-- :func:`crystal_system <matgraphdb.core.nodes.generators.crystal_system>` - A function that generates the crystal systems of a material.
+- :func:`crystal_system <matgraphdb.generators.nodes.crystal_system>` - A function that generates the crystal systems of a material.
 
-- :func:`magnetic_state <matgraphdb.core.nodes.generators.magnetic_state>` - A function that generates the magnetic states of a material.
+- :func:`magnetic_state <matgraphdb.generators.nodes.magnetic_state>` - A function that generates the magnetic states of a material.
 
-- :func:`oxidation_state <matgraphdb.core.nodes.generators.oxidation_state>` - A function that generates the oxidation states of a material.
+- :func:`oxidation_state <matgraphdb.generators.nodes.oxidation_state>` - A function that generates the oxidation states of a material.
 
-- :func:`space_group <matgraphdb.core.nodes.generators.space_group>` - A function that generates the space groups of a material.
+- :func:`space_group <matgraphdb.generators.nodes.space_group>` - A function that generates the space groups of a material.
 
-- :func:`wyckoff <matgraphdb.core.nodes.generators.wyckoff>` - A function that generates the wyckoffs of a material.
+- :func:`wyckoff <matgraphdb.generators.nodes.wyckoff>` - A function that generates the wyckoffs of a material.
+
+- :func:`lattice <matgraphdb.generators.nodes.lattice>` - A function that generates the lattices of a material.
+
+- :func:`material_site <matgraphdb.generators.nodes.material_site>` - A function that generates the sites of a material.
+
+
+
+
 
 
 .. autosummary::
    :toctree: _autosummary
 
-   matgraphdb.core.nodes.generators.element
-   matgraphdb.core.nodes.generators.chemenv
-   matgraphdb.core.nodes.generators.crystal_system
-   matgraphdb.core.nodes.generators.magnetic_state
-   matgraphdb.core.nodes.generators.oxidation_state
-   matgraphdb.core.nodes.generators.space_group
-   matgraphdb.core.nodes.generators.wyckoff
+   matgraphdb.generators.nodes.element
+   matgraphdb.generators.nodes.chemenv
+   matgraphdb.generators.nodes.crystal_system
+   matgraphdb.generators.nodes.magnetic_state
+   matgraphdb.generators.nodes.oxidation_state
+   matgraphdb.generators.nodes.space_group
+   matgraphdb.generators.nodes.wyckoff
+   matgraphdb.generators.nodes.lattice
+   matgraphdb.generators.nodes.material_site
+
