+++ matgraphdb/core/__init__.py
@@ -1,2 +1,2 @@
+from matgraphdb.core.material_store import MaterialStore
 from matgraphdb.core.matgraphdb import MatGraphDB
-from matgraphdb.core.nodes import MaterialStore
