--- matgraphdb/core/datasets/__init__.py
@@ -1,1 +1,0 @@
-from matgraphdb.core.datasets.mp_near_hull import MPNearHull
