+++ matgraphdb/generators/__init__.py
@@ -1,0 +1,2 @@
+from matgraphdb.generators.edges import *
+from matgraphdb.generators.nodes import *
